added multithreading to fdtd, also first commit in a while so probably some other stuff

+++ fdtd/fdtd_io.h
@@ -1,6 +1,6 @@
 #ifndef __FDTD_IO_H__
 #define __FDTD_IO_H__
-void print_as_slices(int time,field H_n,field E_n);
+void print_as_slices(double time,field H_n,field E_n);
 int init_dir(const char* dir_name);
 void dump_data(int time,field H_n,field E_n);
 #endif
